- add services for customers and subscriptions - refactor code structure to use services package
